backend/requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
httpx
websockets
python-multipart
aiomqtt
twilio
python-telegram-bot
supabase
alembic
email-validator
